Refactored a mess of methods back into the base RemoteHttpService class to simplify the various extended classes.
